requirements.txt
google-genai
python-dotenv
requests
streamlit
pandas
Pillow
selenium
webdriver-manager
beautifulsoup4
pytz